feat(minecraft): update to 26.3-snapshot-7
Small changes:
- Introduction of Prediction enum for inventory operations
- Replacement of Swing packet with Punch

See: https://minecraft.wiki/w/Java_Edition_26.3_Snapshot_7

+++ src/mixins/java/org/spongepowered/common/mixin/api/minecraft/world/level/levelgen/DensityFunctionMixin_API.java
@@ -41,7 +41,7 @@ public interface DensityFunctionMixin_API extends org.spongepowered.api.world.ge
 
     // @formatter:off
     @Shadow Interval shadow$range();
-    @Shadow double shadow$compute(final DensityFunction.FunctionContext var1);
+    @Shadow float shadow$compute(final DensityFunction.FunctionContext var1);
     // @formatter:on
 
     @Override
